# gotrip/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GoTrip API",
    "description": "API documentation for GoTrip Travel & Tourism Website",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "fullName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/bookings": {
      "get": {
        "summary": "Get all bookings for the current user",
        "tags": ["Bookings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new booking",
        "tags": ["Bookings"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "destination": {
                    "type": "string"
                  },
                  "travelDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "travelers": {
                    "type": "integer"
                  },
                  "roomType": {
                    "type": "string"
                  },
                  "airportPickup": {
                    "type": "boolean"
                  },
                  "tourGuide": {
                    "type": "boolean"
                  },
                  "insurance": {
                    "type": "boolean"
                  },
                  "specialRequests": {
                    "type": "string"
                  },
                  "totalAmount": {
                    "type": "number"
                  }
                },
                "required": ["fullName", "email", "destination", "travelDate", "totalAmount"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/destinations": {
      "get": {
        "summary": "Get all destinations",
        "tags": ["Destinations"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/destinations/prices/all": {
      "get": {
        "summary": "Get all destination prices with dynamic pricing",
        "tags": ["Destinations"],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/profile": {
      "get": {
        "summary": "Get current user profile",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update user profile",
        "tags": ["Users"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
} 